+++ src/main/java/com/buuz135/industrial/item/ItemArtificalDye.java
@@ -1,0 +1,33 @@
+package com.buuz135.industrial.item;
+
+import net.minecraft.client.renderer.block.model.ModelResourceLocation;
+import net.minecraft.creativetab.CreativeTabs;
+import net.minecraft.item.EnumDyeColor;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.NonNullList;
+import net.minecraft.util.text.TextComponentTranslation;
+import net.minecraftforge.client.model.ModelLoader;
+
+public class ItemArtificalDye extends IFCustomItem {
+    public ItemArtificalDye() {
+        super("artificial_dye");
+        setHasSubtypes(true);
+    }
+
+    @Override
+    public void getSubItems(CreativeTabs tab, NonNullList<ItemStack> subItems) {
+        if (isInCreativeTab(tab))
+            for (int i = 0; i < 16; ++i) subItems.add(new ItemStack(this, 1, i));
+    }
+
+    @Override
+    public void registerRender() {
+        for (int i = 0; i < 16; ++i)
+            ModelLoader.setCustomModelResourceLocation(this, i, new ModelResourceLocation(this.getRegistryName().toString(), "inventory"));
+    }
+
+    @Override
+    public String getItemStackDisplayName(ItemStack stack) {
+        return new TextComponentTranslation("item.fireworksCharge." + EnumDyeColor.byMetadata(stack.getMetadata()).getUnlocalizedName().replaceAll("_", "")).getFormattedText() + " " + super.getItemStackDisplayName(stack);
+    }
+}